Use a redis sorted set for reservation history
We keep a history of reservations, but we typically query them by time
range. Move to a data storage type optimized for that type of query: a
sorted set in redis.

The old data is not removed. That will happen at some point later.
This allows for a smooth transition.

Signed-off-by: Russell Bryant <rbryant@redhat.com>